plugin updates

+++ wp-content/plugins/mailchimp-for-woocommerce/includes/class-mailchimp-woocommerce-service.php
@@ -10,8 +10,6 @@
  */
 class MailChimp_Service extends MailChimp_WooCommerce_Options
 {
-    protected static $pushed_orders = array();
-
     protected $user_email = null;
     protected $previous_email = null;
     protected $force_cart_post = false;
@@ -71,8 +69,8 @@ class MailChimp_Service extends MailChimp_WooCommerce_Options
     {
         if (!mailchimp_is_configured()) return;
 
-        // register this order is already in process..
-        static::$pushed_orders[$order_id] = true;
+        // tell the system the order was brand new - and we don't need to process the order update hook.
+        set_site_transient( "mailchimp_order_created_{$order_id}", true, 20);
 
         // see if we have a session id and a campaign id, also only do this when this user is not the admin.
         $campaign_id = $this->getCampaignTrackingID();
@@ -88,7 +86,7 @@ class MailChimp_Service extends MailChimp_WooCommerce_Options
 
         // queue up the single order to be processed.
         $handler = new MailChimp_WooCommerce_Single_Order($order_id, null, $campaign_id, $landing_site);
-        wp_queue($handler, 60);
+        mailchimp_handle_or_queue($handler, 60);
     }
 
     /**
@@ -99,13 +97,16 @@ class MailChimp_Service extends MailChimp_WooCommerce_Options
     {
         if (!mailchimp_is_configured()) return;
 
-        // register this order is already in process..
-        static::$pushed_orders[$order_id] = true;
+        // if we got a new order hook first - just skip this for now during the 20 second window.
+        if (get_site_transient("mailchimp_order_created_{$order_id}") === true) {
+            return;
+        }
+
         // queue up the single order to be processed.
         $handler = new MailChimp_WooCommerce_Single_Order($order_id, null, null, null);
         $handler->is_update = true;
         $handler->is_admin_save = $is_admin;
-        wp_queue($handler, 90);
+        mailchimp_handle_or_queue($handler, 90);
     }
 
     /**
@@ -117,7 +118,7 @@ class MailChimp_Service extends MailChimp_WooCommerce_Options
 
         $handler = new MailChimp_WooCommerce_Single_Order($order_id, null, null, null);
         $handler->partially_refunded = true;
-        wp_queue($handler);
+        mailchimp_handle_or_queue($handler);
     }
 
     /**
@@ -146,20 +147,30 @@ class MailChimp_Service extends MailChimp_WooCommerce_Options
 
         if (($user_email = $this->getCurrentUserEmail())) {
 
+            // let's skip this right here - no need to go any further.
+            if (mailchimp_email_is_privacy_protected($user_email)) {
+                return !is_null($updated) ? $updated : false;
+            }
+
             $previous = $this->getPreviousEmailFromSession();
 
-            $uid = md5(trim(strtolower($user_email)));
+            $uid = mailchimp_hash_trim_lower($user_email);
+
+            $unique_sid = $this->getUniqueStoreID();
 
             // delete the previous records.
             if (!empty($previous) && $previous !== $user_email) {
 
-                if ($this->api()->deleteCartByID($this->getUniqueStoreID(), $previous_email = md5(trim(strtolower($previous))))) {
+                if ($this->api()->deleteCartByID($unique_sid, $previous_email = mailchimp_hash_trim_lower($previous))) {
                     mailchimp_log('ac.cart_swap', "Deleted cart [$previous] :: ID [$previous_email]");
                 }
 
                 // going to delete the cart because we are switching.
                 $this->deleteCart($previous_email);
             }
+
+            // delete the current cart record if there is one
+            $this->api()->deleteCartByID($unique_sid, $uid);
 
             if ($this->cart && !empty($this->cart)) {
 
@@ -173,7 +184,7 @@ class MailChimp_Service extends MailChimp_WooCommerce_Options
 
                 // fire up the job handler
                 $handler = new MailChimp_WooCommerce_Cart_Update($uid, $user_email, $campaign, $this->cart);
-                wp_queue($handler);
+                mailchimp_handle_or_queue($handler);
             }
 
             return !is_null($updated) ? $updated : true;
@@ -199,7 +210,7 @@ class MailChimp_Service extends MailChimp_WooCommerce_Options
         if (!mailchimp_is_configured()) return;
 
         if ($coupon instanceof WC_Coupon) {
-            wp_queue(new MailChimp_WooCommerce_SingleCoupon($post_id));
+            mailchimp_handle_or_queue(new MailChimp_WooCommerce_SingleCoupon($post_id));
         }
     }
 
@@ -222,9 +233,10 @@ class MailChimp_Service extends MailChimp_WooCommerce_Options
     {
         if (!mailchimp_is_configured()) return;
 
-        if ($post->post_status !== 'auto-draft') {
+        // don't handle any of these statuses because they're not ready for the show
+        if (!in_array($post->post_status, array('trash', 'auto-draft', 'draft', 'pending'))) {
             if ('product' == $post->post_type) {
-                wp_queue(new MailChimp_WooCommerce_Single_Product($post_id), 5);
+                mailchimp_handle_or_queue(new MailChimp_WooCommerce_Single_Product($post_id), 5);
             } elseif ('shop_order' == $post->post_type) {
                 $this->handleOrderStatusChanged($post_id, is_admin());
             }
@@ -240,8 +252,20 @@ class MailChimp_Service extends MailChimp_WooCommerce_Options
 
         switch (get_post_type($post_id)) {
             case 'shop_coupon':
-                mailchimp_get_api()->deletePromoRule(mailchimp_get_store_id(), $post_id);
-                mailchimp_log('promo_code.deleted', "deleted promo code {$post_id}");
+                try {
+                    mailchimp_get_api()->deletePromoRule(mailchimp_get_store_id(), $post_id);
+                    mailchimp_log('promo_code.deleted', "deleted promo code {$post_id}");
+                } catch (\Exception $e) {
+                    mailchimp_error('delete promo code', $e->getMessage());
+                }
+                break;
+            case 'product':
+                try {
+                    mailchimp_get_api()->deleteStoreProduct(mailchimp_get_store_id(), $post_id);
+                    mailchimp_log('product.deleted', "deleted product {$post_id}");
+                } catch (\Exception $e) {
+                    mailchimp_error('delete product', $e->getMessage());
+                }
                 break;
         }
     }
@@ -251,11 +275,20 @@ class MailChimp_Service extends MailChimp_WooCommerce_Options
      */
     public function handlePostRestored($post_id)
     {
-        if (!mailchimp_is_configured()) return;
+        if (!mailchimp_is_configured() || !($post = get_post($post_id))) return;
+
+        // don't handle any of these statuses because they're not ready for the show
+        if (in_array($post->post_status, array('trash', 'auto-draft', 'draft', 'pending'))) {
+            return;
+        }
 
         switch(get_post_type($post_id)) {
             case 'shop_coupon':
                 return $this->handleCouponRestored($post_id);
+                break;
+
+            case 'product':
+                mailchimp_handle_or_queue(new MailChimp_WooCommerce_Single_Product($post_id), 5);
                 break;
         }
     }
@@ -274,7 +307,7 @@ class MailChimp_Service extends MailChimp_WooCommerce_Options
         update_user_meta($user_id, 'mailchimp_woocommerce_is_subscribed', $subscribed);
 
         if ($subscribed) {
-            wp_queue(new MailChimp_WooCommerce_User_Submit($user_id, $subscribed));
+            mailchimp_handle_or_queue(new MailChimp_WooCommerce_User_Submit($user_id, $subscribed));
         }
     }
 
@@ -293,7 +326,7 @@ class MailChimp_Service extends MailChimp_WooCommerce_Options
         if ($is_subscribed === '' || $is_subscribed === null) return;
 
         // only send this update if the user actually has a boolean value.
-        wp_queue(new MailChimp_WooCommerce_User_Submit($user_id, (bool) $is_subscribed, $old_user_data));
+        mailchimp_handle_or_queue(new MailChimp_WooCommerce_User_Submit($user_id, (bool) $is_subscribed, $old_user_data));
     }
 
     /**
@@ -319,7 +352,7 @@ class MailChimp_Service extends MailChimp_WooCommerce_Options
         if (!$this->isAdmin()) return false;
         $this->removePointers(false, true);
         // since the products are all good, let's sync up the orders now.
-        wp_queue(new MailChimp_WooCommerce_Process_Orders());
+        mailchimp_handle_or_queue(new MailChimp_WooCommerce_Process_Orders());
         return true;
     }
